backend/requirements.txt
asyncio==3.4.3
atproto==0.0.56
black==24.10.0
flake8==7.1.1
prometheus_client==0.21.1
motor==3.6.0
python-dotenv==1.0.1
redis==5.2.1
uvicorn==0.34.0
watchdog==6.0.0